requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
mysqlclient>=2.2.0
celery>=5.3.6
redis>=5.0.1
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
django-celery-results>=2.5.1
django-celery-beat>=2.5.0
ultralytics
opencv-python
pyqt5>=5.15.2
lxml>=4.6
